Move python folder hierarchy to core
