+++ src/core/python/utility/pyboost_cv3_converter.cpp
@@ -7,10 +7,13 @@
  *  See the accompanying file LICENSE.txt for the complete license agreement.
  ****************************************************************************/
 
-#include <opencv2/core/core.hpp>
-#if CV_VERSION_EPOCH == 2 || (!defined CV_VERSION_EPOCH && CV_VERSION_MAJOR == 2)
+#include <hipipe/build_config.hpp>
+#if defined HIPIPE_BUILD_PYTHON && defined HIPIPE_BUILD_PYTHON_OPENCV
 
-#include <hipipe/python/utility/pyboost_cv_converter.hpp>
+#include <opencv2/core/core.hpp>
+#if CV_VERSION_MAJOR == 3
+
+#include <hipipe/core/python/utility/pyboost_cv_converter.hpp>
 
 namespace hipipe::python::utility {
 
@@ -60,21 +63,6 @@ catch (const cv::Exception &e) \
     return 0; \
 }
 
-//===================   UTILITY   ==================================================================
-
-static size_t REFCOUNT_OFFSET = (size_t)&(((PyObject*)0)->ob_refcnt) +
-  (0x12345678 != *(const size_t*)"\x78\x56\x34\x12\0\0\0\0\0")*sizeof(int);
-
-static inline PyObject* pyObjectFromRefcount(const int* refcount)
-{
-    return (PyObject*)((size_t)refcount - REFCOUNT_OFFSET);
-}
-
-static inline int* refcountFromPyObject(const PyObject* obj)
-{
-    return (int*)((size_t)obj + REFCOUNT_OFFSET);
-}
-
 //===================   THREADING     ==============================================================
 
 class PyAllowThreads {
@@ -105,56 +93,77 @@ private:
     PyGILState_STATE _state;
 };
 
-//===================   NUMPY ALLOCATOR FOR OPENCV     =============================================
+enum {ARG_NONE = 0, ARG_MAT = 1, ARG_SCALAR = 2};
 
 class NumpyAllocator : public cv::MatAllocator {
 public:
-    void allocate(int dims, const int* sizes, int type, int*& refcount, uchar*& datastart,
-                  uchar*& data, size_t* step)
+    NumpyAllocator()
     {
+        stdAllocator = cv::Mat::getStdAllocator();
+    }
+
+    cv::UMatData* allocate(PyObject* o, int dims, const int* sizes, int type, size_t* step) const
+    {
+        cv::UMatData* u = new cv::UMatData(this);
+        u->data = u->origdata = (uchar*)PyArray_DATA((PyArrayObject*)o);
+        npy_intp* _strides = PyArray_STRIDES((PyArrayObject*)o);
+        for (int i = 0; i < dims - 1; i++) step[i] = (size_t)_strides[i];
+        step[dims - 1] = CV_ELEM_SIZE(type);
+        u->size = sizes[0] * step[0];
+        u->userdata = o;
+        return u;
+    }
+
+    cv::UMatData* allocate(int dims0, const int* sizes, int type, void* data, size_t* step,
+                           int flags, cv::UMatUsageFlags usageFlags) const
+    {
+        if (data != 0) {
+            CV_Error(cv::Error::StsAssert, "The data should normally be NULL!");
+            // probably this is safe to do in such extreme case
+            return stdAllocator->allocate(dims0, sizes, type, data, step, flags, usageFlags);
+        }
         PyEnsureGIL gil;
 
         int depth = CV_MAT_DEPTH(type);
         int cn = CV_MAT_CN(type);
         const int f = (int) (sizeof(size_t) / 8);
-        int typenum = depth == CV_8U ? NPY_UBYTE : depth == CV_8S ? NPY_BYTE :
-                        depth == CV_16U ? NPY_USHORT :
-                        depth == CV_16S ? NPY_SHORT :
-                        depth == CV_32S ? NPY_INT :
-                        depth == CV_32F ? NPY_FLOAT :
-                        depth == CV_64F ? NPY_DOUBLE : f * NPY_ULONGLONG + (f ^ 1) * NPY_UINT;
-        int i;
-        npy_intp _sizes[CV_MAX_DIM + 1];
-        for (i = 0; i < dims; i++) {
-            _sizes[i] = sizes[i];
-        }
+        int typenum = depth == CV_8U ? NPY_UBYTE :
+                      depth == CV_8S ? NPY_BYTE :
+                      depth == CV_16U ? NPY_USHORT :
+                      depth == CV_16S ? NPY_SHORT :
+                      depth == CV_32S ? NPY_INT :
+                      depth == CV_32F ? NPY_FLOAT :
+                      depth == CV_64F ? NPY_DOUBLE :
+                      f * NPY_ULONGLONG + (f ^ 1) * NPY_UINT;
 
-        if (cn > 1) {
-            _sizes[dims++] = cn;
-        }
-
+        int i, dims = dims0;
+        cv::AutoBuffer<npy_intp> _sizes(dims + 1);
+        for (i = 0; i < dims; i++) _sizes[i] = sizes[i];
+        if (cn > 1) _sizes[dims++] = cn;
         PyObject* o = PyArray_SimpleNew(dims, _sizes, typenum);
-
-        if (!o) {
-            CV_Error_(CV_StsError, ("The numpy array of typenum=%d, ndims=%d can not be created",
-                                    typenum, dims));
-        }
-        refcount = refcountFromPyObject(o);
-        npy_intp* _strides = PyArray_STRIDES((PyArrayObject*)o);
-        for (i = 0; i < dims - (cn > 1); i++) {
-            step[i] = (size_t)_strides[i];
-        }
-        datastart = data = (uchar*)PyArray_DATA((PyArrayObject*)o);
+        if (!o)
+            CV_Error_(
+              cv::Error::StsError,
+              ("The numpy array of typenum=%d, ndims=%d can not be created", typenum, dims));
+        return allocate(o, dims0, sizes, type, step);
     }
 
-    void deallocate(int* refcount, uchar*, uchar*)
+    bool allocate(cv::UMatData* u, int accessFlags, cv::UMatUsageFlags usageFlags) const
     {
-        PyEnsureGIL gil;
-        if (!refcount) return;
-        PyObject* o = pyObjectFromRefcount(refcount);
-        Py_INCREF(o);
-        Py_DECREF(o);
+        return stdAllocator->allocate(u, accessFlags, usageFlags);
     }
+
+    void deallocate(cv::UMatData* u) const
+    {
+        if (u) {
+            PyEnsureGIL gil;
+            PyObject* o = (PyObject*)u->userdata;
+            Py_XDECREF(o);
+            delete u;
+        }
+    }
+
+    const cv::MatAllocator* stdAllocator;
 };
 
 //===================   ALLOCATOR INITIALIZTION   ==================================================
@@ -167,33 +176,34 @@ PyObject* fromMatToNDArray(const cv::Mat& m)
 {
     if (!m.data) Py_RETURN_NONE;
     cv::Mat temp, *p = (cv::Mat *)&m;
-    if (!p->refcount || p->allocator != &g_numpyAllocator) {
+    if (!p->u || p->allocator != &g_numpyAllocator) {
         temp.allocator = &g_numpyAllocator;
         ERRWRAP2(m.copyTo(temp));
         p = &temp;
     }
-    p->addref();
-    return pyObjectFromRefcount(p->refcount);
+    PyObject* o = (PyObject*)p->u->userdata;
+    Py_INCREF(o);
+    return o;
 }
 
 cv::Mat fromNDArrayToMat(PyObject* o) {
     cv::Mat m;
+    bool allowND = true;
     if (!PyArray_Check(o)) {
         failmsg("argument is not a numpy array");
         if (!m.data) m.allocator = &g_numpyAllocator;
     } else {
-        PyArrayObject* oarr = (PyArrayObject*) o;
+        PyArrayObject* oarr = (PyArrayObject*)o;
 
         bool needcopy = false, needcast = false;
         int typenum = PyArray_TYPE(oarr), new_typenum = typenum;
-        int type = typenum == NPY_UBYTE ? CV_8U :
-                   typenum == NPY_BYTE ? CV_8S :
-                   typenum == NPY_USHORT ? CV_16U :
-                   typenum == NPY_SHORT ? CV_16S :
-                   typenum == NPY_INT ? CV_32S :
-                   typenum == NPY_INT32 ? CV_32S :
-                   typenum == NPY_FLOAT ? CV_32F :
-                   typenum == NPY_DOUBLE ? CV_64F : -1;
+        int type = typenum == NPY_UBYTE ? CV_8U : typenum == NPY_BYTE ? CV_8S :
+                    typenum == NPY_USHORT ? CV_16U :
+                    typenum == NPY_SHORT ? CV_16S :
+                    typenum == NPY_INT ? CV_32S :
+                    typenum == NPY_INT32 ? CV_32S :
+                    typenum == NPY_FLOAT ? CV_32F :
+                    typenum == NPY_DOUBLE ? CV_64F : -1;
 
         if (type < 0) {
             if (typenum == NPY_INT64 || typenum == NPY_UINT64 || type == NPY_LONG) {
@@ -217,6 +227,7 @@ cv::Mat fromNDArrayToMat(PyObject* o) {
             if (!m.data) m.allocator = &g_numpyAllocator;
             return m;
         }
+
         int size[CV_MAX_DIM + 1];
         size_t step[CV_MAX_DIM + 1];
         size_t elemsize = CV_ELEM_SIZE1(type);
@@ -226,7 +237,8 @@ cv::Mat fromNDArrayToMat(PyObject* o) {
 
         for (int i = ndims - 1; i >= 0 && !needcopy; i--) {
             // these checks handle cases of
-            //  a) multi-dimensional (ndims > 2) arrays, as well as simpler 1- and 2-dimensional cases
+            //  a) multi-dimensional (ndims > 2) arrays, as well as simpler 1- and 2-dimensional
+            //  cases
             //  b) transposed arrays, where _strides[] elements go in non-descending order
             //  c) flipped arrays, where some of _strides[] elements are negative
             if ((i == ndims - 1 && (size_t)_strides[i] != elemsize) ||
@@ -234,16 +246,16 @@ cv::Mat fromNDArrayToMat(PyObject* o) {
                 needcopy = true;
         }
 
-        if (ismultichannel && _strides[1] != (npy_intp) elemsize * _sizes[2]) {
+        if (ismultichannel && _strides[1] != (npy_intp)elemsize * _sizes[2]) {
             needcopy = true;
         }
         if (needcopy) {
             if (needcast) {
                 o = PyArray_Cast(oarr, new_typenum);
-                oarr = (PyArrayObject*) o;
+                oarr = (PyArrayObject*)o;
             } else {
                 oarr = PyArray_GETCONTIGUOUS(oarr);
-                o = (PyObject*) oarr;
+                o = (PyObject*)oarr;
             }
             _strides = PyArray_STRIDES(oarr);
         }
@@ -259,18 +271,21 @@ cv::Mat fromNDArrayToMat(PyObject* o) {
             step[ndims] = elemsize;
             ndims++;
         }
+
         if (ismultichannel) {
             ndims--;
             type |= CV_MAKETYPE(0, size[2]);
         }
 
-        m = cv::Mat(ndims, size, type, PyArray_DATA(oarr), step);
+        if (ndims > 2 && !allowND) {
+            failmsg("%s has more than 2 dimensions");
+        } else {
+            m = cv::Mat(ndims, size, type, PyArray_DATA(oarr), step);
+            m.u = g_numpyAllocator.allocate(o, ndims, size, type, step);
+            m.addref();
 
-        if (m.data) {
-            m.refcount = refcountFromPyObject(o);
             if (!needcopy) {
-                m.addref();  // protect the original numpy array from deallocation
-                             // (since Mat destructor will decrement the reference counter)
+                Py_INCREF(o);
             }
         }
         m.allocator = &g_numpyAllocator;
@@ -280,33 +295,34 @@ cv::Mat fromNDArrayToMat(PyObject* o) {
 
 //===================   BOOST CONVERTERS     =======================================================
 
-PyObject* matToNDArrayBoostConverter::convert(cv::Mat const& m) {
-    if( !m.data )
-        Py_RETURN_NONE;
-    cv::Mat temp, *p = (cv::Mat*)&m;
-    if(!p->refcount || p->allocator != &g_numpyAllocator)
-    {
+PyObject* matToNDArrayBoostConverter::convert(cv::Mat const& m)
+{
+    if (!m.data) Py_RETURN_NONE;
+    cv::Mat temp, *p = (cv::Mat *)&m;
+    if (!p->u || p->allocator != &g_numpyAllocator) {
         temp.allocator = &g_numpyAllocator;
         ERRWRAP2(m.copyTo(temp));
         p = &temp;
     }
-    p->addref();
-    return pyObjectFromRefcount(p->refcount);
+    PyObject* o = (PyObject*)p->u->userdata;
+    Py_INCREF(o);
+    return o;
 }
 
-matFromNDArrayBoostConverter::matFromNDArrayBoostConverter() {
-    boost::python::converter::registry::push_back(matFromNDArrayBoostConverter::convertible,
-            matFromNDArrayBoostConverter::construct,
-            boost::python::type_id<cv::Mat>());
+matFromNDArrayBoostConverter::matFromNDArrayBoostConverter()
+{
+    boost::python::converter::registry::push_back(convertible, construct,
+                                                  boost::python::type_id<cv::Mat>());
 }
 
-/// @brief Check if PyObject is an array and can be converted to OpenCV matrix.
-void* matFromNDArrayBoostConverter::convertible(PyObject* object) {
+// check if PyObject is an array and can be converted to OpenCV matrix.
+void* matFromNDArrayBoostConverter::convertible(PyObject* object)
+{
     if (!PyArray_Check(object)) return NULL;
 #ifndef CV_MAX_DIM
     const int CV_MAX_DIM = 32;
 #endif
-    PyArrayObject* oarr = (PyArrayObject*) object;
+    PyArrayObject* oarr = (PyArrayObject*)object;
 
     int typenum = PyArray_TYPE(oarr);
     if (typenum != NPY_INT64 && typenum != NPY_UINT64 && typenum != NPY_LONG &&
@@ -317,17 +333,14 @@ void* matFromNDArrayBoostConverter::convertible(PyObject* object) {
     }
     int ndims = PyArray_NDIM(oarr);  // data type not supported
 
-    if (ndims >= CV_MAX_DIM) {
-        return NULL;  // too many dimensions
-    }
-
+    if (ndims >= CV_MAX_DIM) return NULL;  // too many dimensions
     return object;
 }
 
+// construct a Mat from an NDArray object.
 void matFromNDArrayBoostConverter::construct(
   PyObject* object, boost::python::converter::rvalue_from_python_stage1_data* data)
 {
-
     namespace python = boost::python;
     // Object is a borrowed reference, so create a handle indicting it is
     // borrowed for proper reference counting.
@@ -346,14 +359,13 @@ void matFromNDArrayBoostConverter::construct(
 
     bool needcopy = false, needcast = false;
     int typenum = PyArray_TYPE(oarr), new_typenum = typenum;
-    int type = typenum == NPY_UBYTE ? CV_8U :
-               typenum == NPY_BYTE ? CV_8S :
-               typenum == NPY_USHORT ? CV_16U :
-               typenum == NPY_SHORT ? CV_16S :
-               typenum == NPY_INT ? CV_32S :
-               typenum == NPY_INT32 ? CV_32S :
-               typenum == NPY_FLOAT ? CV_32F :
-               typenum == NPY_DOUBLE ? CV_64F : -1;
+    int type = typenum == NPY_UBYTE ? CV_8U : typenum == NPY_BYTE ? CV_8S :
+                typenum == NPY_USHORT ? CV_16U :
+                typenum == NPY_SHORT ? CV_16S :
+                typenum == NPY_INT ? CV_32S :
+                typenum == NPY_INT32 ? CV_32S :
+                typenum == NPY_FLOAT ? CV_32F :
+                typenum == NPY_DOUBLE ? CV_64F : -1;
 
     if (type < 0) {
         needcopy = needcast = true;
@@ -365,7 +377,6 @@ void matFromNDArrayBoostConverter::construct(
     const int CV_MAX_DIM = 32;
 #endif
     int ndims = PyArray_NDIM(oarr);
-
     int size[CV_MAX_DIM + 1];
     size_t step[CV_MAX_DIM + 1];
     size_t elemsize = CV_ELEM_SIZE1(type);
@@ -389,10 +400,10 @@ void matFromNDArrayBoostConverter::construct(
     if (needcopy) {
         if (needcast) {
             object = PyArray_Cast(oarr, new_typenum);
-            oarr = (PyArrayObject*) object;
+            oarr = (PyArrayObject*)object;
         } else {
             oarr = PyArray_GETCONTIGUOUS(oarr);
-            object = (PyObject*) oarr;
+            object = (PyObject*)oarr;
         }
         _strides = PyArray_STRIDES(oarr);
     }
@@ -418,17 +429,13 @@ void matFromNDArrayBoostConverter::construct(
     }
 
     cv::Mat* m = new (storage) cv::Mat(ndims, size, type, PyArray_DATA(oarr), step);
-    if (m->data) {
-        m->refcount = refcountFromPyObject(object);
-        if (!needcopy) {
-            m->addref();  // protect the original numpy array from deallocation
-                          // (since Mat destructor will decrement the reference counter)
-        }
-    };
-
+    m->u = g_numpyAllocator.allocate(object, ndims, size, type, step);
     m->allocator = &g_numpyAllocator;
+    m->addref();
     data->convertible = storage;
 }
 
 }  // namespace hipipe::python::utility
-#endif
+
+#endif // CV_VERSION_MAJOR == 3
+#endif // defined HIPIPE_BUILD_PYTHON && defined HIPIPE_BUILD_PYTHON_OPENCV
